Code cleanup: unused imports, loading skeleton, PII logging, type safety
Round 3 fixes (10 low-severity cleanup items):
- Remove unused InlineStack import from PauseModal
- Align ProfileBlock types with canonical types.ts (union types, FREQUENCY_LABELS)
- Add SkeletonTextBlock loading state to SubscriptionPage
- Add "Try Again" retry button on error state in SubscriptionPage
- Replace hardcoded fallback time slots with "no slots available" message
- Use parseFloat instead of parseInt for discount values (supports 2.5%)
- Remove unused locale keys (title, shipping) from en.default.json
- Reduce verbose PII logging in orders webhook (remove full GraphQL dumps)
- Replace (api as any) with typed GraphQLResult/ExtensionData interfaces
- Remove customer email from API GET response (unnecessary exposure)

Skipped #27 (shared sessionStorage key is intentional) and #28 (cancelled
subscriptions Past section is a feature enhancement, not cleanup.

EOF
)

+++ extensions/pickup-scheduler-cart/assets/subscribe-save.js
@@ -448,7 +448,7 @@
       return {
         type: checked.value,
         frequency: checked.dataset.frequency || null,
-        discount: parseInt(checked.dataset.discount) || 0
+        discount: parseFloat(checked.dataset.discount) || 0
       };
     }
   }
@@ -467,9 +467,9 @@
     }
 
     // Get settings from the embed element
-    const weeklyDiscount = parseInt(embed.dataset.weeklyDiscount) || 10;
-    const biweeklyDiscount = parseInt(embed.dataset.biweeklyDiscount) || 5;
-    const triweeklyDiscount = parseInt(embed.dataset.triweeklyDiscount) || 0;
+    const weeklyDiscount = parseFloat(embed.dataset.weeklyDiscount) || 10;
+    const biweeklyDiscount = parseFloat(embed.dataset.biweeklyDiscount) || 5;
+    const triweeklyDiscount = parseFloat(embed.dataset.triweeklyDiscount) || 0;
     const shopDomain = embed.dataset.shop || window.Shopify?.shop || '';
 
     // Create the widget
